So, it's been a while. I made a bunch of changes in a short time and lost track of when the changes were made. Here's a very brief, inadequate description:
Added a lot of little functions to accomplish various tasks...that's why most of the changes are in the utility folder. Many of these may not remain, but they were useful at the time. Mostly just little snippets to accomplish tasks...mostly in vmware, but using data from other systems.

I also started the process of adding proper documentation to my functions.